URL validation fix

+++ main.go
@@ -4,6 +4,7 @@ import (
 	tb "gopkg.in/tucnak/telebot.v2"
 	"io/ioutil"
 	"log"
+	"net/http"
 	"os"
 	"os/exec"
 	"time"
@@ -25,36 +26,45 @@ func SendMP3(){
 	log.Println("Bot is running")
 
 	bot.Handle("/start", func(m *tb.Message) {
-		bot.Send(m.Sender,"Youtube2MP3 botuna hoşgeldiniz :)\nİndirmek istediğiniz YouTube videosunun linkini mesaj olarak göndermeniz yeterlidir.\n ")
+		bot.Send(m.Sender,"Youtube2MP3 botuna hoşgeldiniz :)\n" +
+			"İndirmek istediğiniz YouTube videosunun linkini mesaj olarak göndermeniz yeterlidir.\n" +
+			"Katkıda bulunmak için : github.com/msrexe/yt2mp3-telegram-bot\n" +
+			"@msrexe")
 	})
 	bot.Handle(tb.OnText, func(m *tb.Message) {
 		if urlCheck(m.Text) {
 			bot.Send(m.Sender,"İsteğiniz işleniyor...")
-			bot.Send(m.Sender,"Dosya hazırlanıyor (Bu işlem biraz zaman alabilir)...")
 			dir,_ := ioutil.TempDir("/","prefix")
-			downloadMP3(m.Text)
-			bot.Send(m.Sender,"Az kaldı ...")
-			mp3 := &tb.Audio{
-				File: tb.FromDisk("file.mp3"),
+			err := downloadMP3(m.Text)
+			if err == nil {
+				bot.Send(m.Sender,"Dosya hazırlanıyor (Bu işlem biraz zaman alabilir)...")
+				mp3 := &tb.Audio{
+					File: tb.FromDisk("file.mp3"),
+				}
+				bot.Send(m.Sender,mp3)
+				os.RemoveAll(dir)
+			}else{
+				bot.Send(m.Sender,"Video bulunamadı. Kaldırılmış veya hatalı URL olabilir!!")
 			}
-			bot.Send(m.Sender,mp3)
-			os.RemoveAll(dir)
 		}else{
 			bot.Send(m.Sender,"Geçersiz URL!! \n Örnek istek => 'https://www.youtube.com/watch?v=jHjFxJVeCQs'")
 		}
 	})
-
 	bot.Start()
 }
 
-func downloadMP3(url string){
+func downloadMP3(url string) error{
 	command1 := exec.Command("cd","/utils")
 	command2 := exec.Command("youtube-dl","-x","--audio-format","mp3","-o","/file.mp4",url)
-	log.Println(command2)
 	command1.Run()
-	command2.Run()
+	_,err := command2.Output()
+	return err
 }
 
-func urlCheck(url string) bool {
+func urlCheck(urlString string) bool {
+	_ ,err := http.Get(urlString)
+	if err != nil {
+		return false
+	}
 	return true
 }
